requirements.txt
polars
sentence-transformers
scikit-learn
matplotlib
seaborn
networkx
nltk
python-dotenv
google-generativeai
requests
openai
numpy
pandas
torch
transformers
thefuzz[speedup]
bs4
pyarrow
aiohttp
umap_learn
pyvis